executor: remove the responsibility to mount the DirPoolVolume
It is hard to ensure that every executor routine will see
a consistent view of the mount status of DirPoolVolume.
So we just remove the responsibility to mount it.
I add an extra config "dir_pool_volume_mountpoint",
so heketi administrator should be responsible to
ensure the mountpoint exists, or heketi will not work.

Signed-off-by: guihecheng <guihecheng@cmiot.chinamobile.com>